Added Nico and made a "Hello World"
Nico seemed like it could be a helpful tool for creating the application window/ the rest of the game, so i added it in the source file. Inside the Nico folder stranger_jam application file is in there. Run it and it should open a window saying "Hello World". If you don't want this let me know.