Created Music. Read Description.
Since we are trying to make a puzzle game I used a scale known for being more mysterious, but if you want a different mood, let me know. This is a sample, not the full version. Didn't want to overextend on something if we didn't like the feel/sound of. Had issues uploading, so I had to zip it.